refactor(site): consolidate src/utils into src/util
